backend-api/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sqlalchemy
apscheduler
tzdata
